requirements.txt
mlx>=0.20.0
torch
numpy
diffusers>=0.25.0
transformers
huggingface_hub
Pillow
safetensors
accelerate
hf_transfer
tqdm
gradio
